@ Consolidate EAR tests into shared GlassFishDeployEarTest
GlassFishManagedDeployEarTest already covered GlassFishRestDeployEarTest
via the Derby-aware greeterImplementationBasedOnDerbyEnabled() factory
method. Removed managed/ and remote/ sub-packages — all tests now live
in the single ...test package.

@
